startup should now start the webserver.
I removed the CGI stuff since i dont think we are using that, but it
might be needed to load the first webpage in. we can investigate this
afternoon. Jon to do: change the SSI file names in the array in
startup.c and add the fileread stuff for javascript.
I left in io_init(). you can call the initialization of your stuff that
or change that in startup.c